+++ lexicmap/cmd/lib-chaining2.go
@@ -41,7 +41,7 @@ var DefaultChaining2Options = Chaining2Options{
 	MinScore:    50,
 	MinAlignLen: 50,
 
-	MaxDistance: 50,
+	MaxDistance: 100,
 	Band:        100,
 }
 
@@ -57,7 +57,7 @@ type Chainer2 struct {
 	// maxscoresIdxs []int
 	maxscoresIdxs []uint64 // pack score (32bit) and index (32bit) to save ram.
 
-	bounds []int32 // 4 * chains, can't limit this or it will miss some cases.
+	// bounds []int32 // 4 * chains, can't limit this or it will miss some cases.
 }
 
 // NewChainer creates a new chainer.
@@ -68,7 +68,7 @@ func NewChainer2(options *Chaining2Options) *Chainer2 {
 		// scores:        make([]int, 0, 10240),
 		// maxscores:     make([]int, 0, 51200),
 		maxscoresIdxs: make([]uint64, 0, 51200),
-		bounds:        make([]int32, 128),
+		// bounds:        make([]int32, 128),
 	}
 	return c
 }
@@ -97,13 +97,17 @@ var poolChain2 = &sync.Pool{New: func() interface{} {
 type Chain2Result struct {
 	NAnchors int // The number of substrings
 
+	AlignedFraction float64
+
 	MatchedBases  int     // The number of matched bases.
 	AlignedBasesQ int     // The number of aligned bases in Query sequence
 	AlignedBasesT int     // The number of aligned bases in Subject sequence
-	Pident        float64 // percentage of identity
+	PIdent        float64 // percentage of identity
 
 	QBegin, QEnd int // Query begin/end position (0-based)
 	TBegin, TEnd int // Target begin/end position (0-based)
+
+	TSeq []byte // target seq
 }
 
 // Reset resets a Chain2Result
@@ -140,7 +144,7 @@ func (ce *Chainer2) Chain(subs *[]*SubstrPair) (*[]*Chain2Result, int, int, int,
 			path.QBegin, path.QEnd = qb, qe
 			path.TBegin, path.TEnd = tb, te
 			path.MatchedBases = slen
-			path.Pident = 100
+			path.PIdent = 100
 			path.AlignedBasesQ = slen
 			path.NAnchors++
 			*paths = append(*paths, path)
@@ -205,7 +209,7 @@ func (ce *Chainer2) Chain(subs *[]*SubstrPair) (*[]*Chain2Result, int, int, int,
 				continue
 			}
 
-			d = distance2(a, b)
+			d = distance2(a, b) + float64(b.Len)
 			if d > maxDistance { // limit the distance. necessary?
 				continue
 			}
@@ -260,7 +264,7 @@ func (ce *Chainer2) Chain(subs *[]*SubstrPair) (*[]*Chain2Result, int, int, int,
 	*paths = (*paths)[:0]
 
 	var nMatchedBases, nAlignedBasesQ, nAlignedBasesT int
-	ce.bounds = ce.bounds[:0]
+	// ce.bounds = ce.bounds[:0]
 
 	_, qB, qE, tB, tE := chainARegion(
 		subs,
@@ -275,7 +279,7 @@ func (ce *Chainer2) Chain(subs *[]*SubstrPair) (*[]*Chain2Result, int, int, int,
 		&nAlignedBasesQ,
 		&nAlignedBasesT,
 		Mi,
-		&ce.bounds,
+		nil, // &ce.bounds,
 	)
 
 	return paths, nMatchedBases, nAlignedBasesQ, nAlignedBasesT, qB, qE, tB, tE
@@ -333,8 +337,8 @@ func chainARegion(subs *[]*SubstrPair, // a region of the subs
 	var sub *SubstrPair
 	var beginOfNextAnchor int
 	var pident float64
-	var overlapped bool
-	var nb, bi, bj int // index of bounds
+	// var overlapped bool
+	// var nb, bi, bj int // index of bounds
 	firstAnchorOfAChain := true
 	path := poolChain2.Get().(*Chain2Result)
 	path.Reset()
@@ -359,47 +363,48 @@ func chainARegion(subs *[]*SubstrPair, // a region of the subs
 		// o-------------------- Ref
 		//
 		sub = (*subs)[i]
-		overlapped = false
-		nb = len(*bounds) >> 2 // len(bounds) / 4
-		for bi = 0; bi < nb; bi++ {
-			bj = bi << 2
-			if !((sub.QBegin > (*bounds)[bj+1] && sub.TBegin > (*bounds)[bj+3]) || // top right
-				(sub.QBegin+int32(sub.Len)-1 < (*bounds)[bj] && sub.TBegin+int32(sub.Len)-1 < (*bounds)[bj+2])) { // bottom left
-				overlapped = true
-				break
-			}
-		}
 
-		if overlapped {
-			// fmt.Printf("  %d (%s) is overlapped previous chain, j=%d\n", i, *sub, j)
+		// overlapped = false
+		// nb = len(*bounds) >> 2 // len(bounds) / 4
+		// for bi = 0; bi < nb; bi++ {
+		// 	bj = bi << 2
+		// 	if !((sub.QBegin > (*bounds)[bj+1] && sub.TBegin > (*bounds)[bj+3]) || // top right
+		// 		(sub.QBegin+int32(sub.Len)-1 < (*bounds)[bj] && sub.TBegin+int32(sub.Len)-1 < (*bounds)[bj+2])) { // bottom left
+		// 		overlapped = true
+		// 		break
+		// 	}
+		// }
 
-			// can not continue here, must check if i == j
+		// if overlapped {
+		// 	// fmt.Printf("  %d (%s) is overlapped previous chain, j=%d\n", i, *sub, j)
+
+		// 	// can not continue here, must check if i == j
+		// } else {
+		// path.Chain = append(path.Chain, i+offset) // record the seed
+		path.NAnchors++
+
+		// fmt.Printf(" AAADDD %d (%s). firstAnchorOfAChain: %v\n", i, *sub, firstAnchorOfAChain)
+
+		if firstAnchorOfAChain {
+			// fmt.Printf(" record bound beginning with: %s\n", sub)
+			firstAnchorOfAChain = false
+
+			qe = int32(sub.QBegin) + int32(sub.Len) - 1   // end
+			te = int32(sub.TBegin) + int32(sub.Len) - 1   // end
+			qb, tb = int32(sub.QBegin), int32(sub.TBegin) // in case there's only one anchor
+
+			nMatchedBases += int(sub.Len)
 		} else {
-			// path.Chain = append(path.Chain, i+offset) // record the seed
-			path.NAnchors++
+			qb, tb = int32(sub.QBegin), int32(sub.TBegin) // begin
 
-			// fmt.Printf(" AAADDD %d (%s). firstAnchorOfAChain: %v\n", i, *sub, firstAnchorOfAChain)
-
-			if firstAnchorOfAChain {
-				// fmt.Printf(" record bound beginning with: %s\n", sub)
-				firstAnchorOfAChain = false
-
-				qe = int32(sub.QBegin) + int32(sub.Len) - 1   // end
-				te = int32(sub.TBegin) + int32(sub.Len) - 1   // end
-				qb, tb = int32(sub.QBegin), int32(sub.TBegin) // in case there's only one anchor
-
-				nMatchedBases += int(sub.Len)
+			if int(sub.QBegin)+int(sub.Len)-1 >= beginOfNextAnchor {
+				nMatchedBases += beginOfNextAnchor - int(sub.QBegin)
 			} else {
-				qb, tb = int32(sub.QBegin), int32(sub.TBegin) // begin
-
-				if int(sub.QBegin)+int(sub.Len)-1 >= beginOfNextAnchor {
-					nMatchedBases += beginOfNextAnchor - int(sub.QBegin)
-				} else {
-					nMatchedBases += int(sub.Len)
-				}
+				nMatchedBases += int(sub.Len)
 			}
-			beginOfNextAnchor = int(sub.QBegin)
 		}
+		beginOfNextAnchor = int(sub.QBegin)
+		// }
 
 		if i == j { // the path starts here
 			if firstAnchorOfAChain { // sadly, there's no anchor added.
@@ -428,7 +433,7 @@ func chainARegion(subs *[]*SubstrPair, // a region of the subs
 			path.AlignedBasesQ = nAlignedBasesQ
 			path.AlignedBasesT = nAlignedBasesT
 			path.MatchedBases = nMatchedBases
-			path.Pident = pident
+			path.PIdent = pident
 			path.QBegin, path.QEnd = int(qb), int(qe)
 			path.TBegin, path.TEnd = int(tb), int(te)
 			*paths = append(*paths, path)
@@ -467,7 +472,7 @@ func chainARegion(subs *[]*SubstrPair, // a region of the subs
 					path.AlignedBasesQ = nAlignedBasesQ
 					path.AlignedBasesT = nAlignedBasesT
 					path.MatchedBases = nMatchedBases
-					path.Pident = pident
+					path.PIdent = pident
 					path.QBegin, path.QEnd = int(qb), int(qe)
 					path.TBegin, path.TEnd = int(tb), int(te)
 					*paths = append(*paths, path)
@@ -483,10 +488,10 @@ func chainARegion(subs *[]*SubstrPair, // a region of the subs
 		}
 	}
 
-	*bounds = append(*bounds, qb)
-	*bounds = append(*bounds, qe)
-	*bounds = append(*bounds, tb)
-	*bounds = append(*bounds, te)
+	// *bounds = append(*bounds, qb)
+	// *bounds = append(*bounds, qe)
+	// *bounds = append(*bounds, tb)
+	// *bounds = append(*bounds, te)
 
 	// initialize the boundary
 	qB, qE = int(qb), int(qe)
@@ -579,8 +584,8 @@ func gap2(a, b *SubstrPair) float64 {
 }
 
 func gapScore2(gap float64) float64 {
-	if gap == 0 {
-		return 0
-	}
+	// if gap == 0 {
+	// 	return 0
+	// }
 	return 0.1 * gap
 }